[FEATURE] Introduce native JSON database field type support
The patch adds native JSON database field type.

This allows to define a JSON database field in extension
`ext_tables.sql` files, example:

CREATE TABLE tx_myextension_domain_model_book (
    title varchar(200) DEFAULT '',
    contents json
);

The methods `insert()`, `bulkInsert()` and `update()` in
class `\TYPO3\CMS\Core\Database\Connection` now
convert provided data values to their database counterpart
as define by doctrine/dbal. For json fields, values can be
provides as array and will be automatically json_encode()'ed.

The patch:
* Ensures php to database value conversion in fascade connection
  methods `insert()`, 'bulkInsert()` and `update()`
* Map JSON type from ext_tables.sql to doctrine/dbal JSON Type
* Remove platform specific automatic type determination in core
  DataHandler due to generic handling in Connection fascade class
* doctrine/dbal DateTimeType, DateType and TimeType are overriden
  to allow string as php value for converting to database

Resolves: #99062
Releases: main
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76553
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>

+++ typo3/sysext/install/Classes/Service/ClearCacheService.php
@@ -68,7 +68,9 @@ class ClearCacheService
         $connectionPool = GeneralUtility::makeInstance(ConnectionPool::class);
         foreach (self::legacyDatabaseCacheTables as $tableName) {
             $connection = $connectionPool->getConnectionForTable($tableName);
-            $connection->truncate($tableName);
+            if ($connection->createSchemaManager()->tablesExist($tableName)) {
+                $connection->truncate($tableName);
+            }
         }
 
         // Flush all caches defined in TYPO3_CONF_VARS, but not the ones defined by extensions in ext_localconf.php
